Get informatino from canvas
