+++ frontend/app/pong/[id]/page.tsx
@@ -1,104 +1,40 @@
 "use client";
-import { Button } from "@/components/ui/button";
-import { io, Socket } from "socket.io-client";
-import { useEffect, useRef, useState } from "react";
-import { PongGame } from "./PongGame";
-import { TARGET_FRAME_MS } from "./const";
+import { useCallback, useEffect, useRef, useState } from "react";
+import PongInformationBoard from "./PongInformationBoard";
+import PongBoard from "./PongBoard";
 
 export default function Page({ params: { id } }: { params: { id: string } }) {
-  const canvasRef = useRef<HTMLCanvasElement>(null);
-  const [socket, setSocket] = useState<Socket>(undefined!);
-  // todo: useRef vs useState
-  const game = useRef<PongGame>(new PongGame());
+  const [fps, setFps] = useState<number>(0);
+  const [speed, setSpeed] = useState<number>(0);
+  const [player1Position, setPlayer1Position] = useState<number>(0);
+  const [player2Position, setPlayer2Position] = useState<number>(0);
 
-  useEffect(() => {
-    // > If the contextType doesn't match a possible drawing context, or differs from the first contextType requested, null is returned."
-    // from https://developer.mozilla.org/en-US/docs/Web/API/HTMLCanvasElement/getContext
-    const ctx = canvasRef.current?.getContext("2d");
-    if (!ctx) {
-      console.warn("2d canvas is not supported or there is a bug");
-      return;
-    }
-    const socket = io(process.env.NEXT_PUBLIC_WEB_URL as string);
-    setSocket(socket);
-
-    game.current.setup_canvas(ctx, socket);
-    game.current.draw_canvas();
-    const intervalId = setInterval(game.current.update, TARGET_FRAME_MS);
-
-    const handleKeyUp = (event: KeyboardEvent) => {
-      game.current.keypress[event.key] = false;
-    };
-    const handleKeyDown = (event: KeyboardEvent) => {
-      game.current.keypress[event.key] = true;
-    };
-
-    document.addEventListener("keydown", handleKeyDown);
-
-    document.addEventListener("keyup", handleKeyUp);
-
-    socket.on("connect", () => {
-      console.log(`Connected: ${socket.id}`);
-
-      socket.emit("join", id);
-    });
-
-    socket.on("start", (data) => {
-      console.log(`Start: ${JSON.stringify(data)}`);
-      game.current.start(data);
-    });
-
-    socket.on("right", () => {
-      game.current.player2.clear(game.current.ctx);
-      game.current.player2.move_left();
-      game.current.player2.draw(game.current.ctx);
-    });
-
-    socket.on("left", () => {
-      game.current.player2.clear(game.current.ctx);
-      game.current.player2.move_right();
-      game.current.player2.draw(game.current.ctx);
-    });
-
-    socket.on("bounce", () => {
-      game.current.ball.bounce_off_paddle(game.current.player2);
-    });
-
-    socket.on("collide", () => {
-      game.current.ball.reset();
-      game.current.score.player1++;
-    });
-
-    return () => {
-      socket.disconnect();
-      document.removeEventListener("keydown", handleKeyDown);
-      document.removeEventListener("keyup", handleKeyUp);
-      clearInterval(intervalId);
-    };
-  }, [id]);
-
-  const start = () => {
-    game.current.start();
-    socket.emit("start", {
-      vx: -game.current.ball.vx,
-      vy: -game.current.ball.vy,
-    });
-  };
+  const memoizedSetFps = useCallback((fps: number) => setFps(fps), []);
+  const memoizedSetSpeed = useCallback((speed: number) => setSpeed(speed), []);
+  const memoizedSetPlayer1Position = useCallback(
+    (player1Position: number) => setPlayer1Position(player1Position),
+    [],
+  );
+  const memoizedSetPlayer2Position = useCallback(
+    (player2Position: number) => setPlayer2Position(player2Position),
+    [],
+  );
 
   return (
     <>
-      {/* <PongInformationBoard {} /> */}
-      <div>
-        <Button onClick={start}>Start</Button>
-        <Button onClick={game.current.switch_battle_mode}>Battle</Button>
-        <Button onClick={game.current.switch_practice_mode}>Practice</Button>
-      </div>
-      <canvas
-        ref={canvasRef}
-        width="256"
-        height="512"
-        className="border w-[256px] h-[512px]"
-      ></canvas>
+      <PongInformationBoard
+        fps={fps}
+        speed={speed}
+        player1Position={player1Position}
+        player2Position={player2Position}
+      />
+      <PongBoard
+        id={id}
+        setFps={memoizedSetFps}
+        setSpeed={memoizedSetSpeed}
+        setPlayer1Position={memoizedSetPlayer1Position}
+        setPlayer2Position={memoizedSetPlayer2Position}
+      />
     </>
   );
 }
